aws/backend/requirements.txt
django==5.0.3
djangorestframework==3.15.1
mangum==0.17.0
psycopg2-binary==2.9.9
boto3==1.34.0
PyJWT==2.8.0
cryptography==42.0.0
requests==2.31.0
django-cors-headers==4.3.1
openai==1.35.0
